rename and refactor ruby transform tests
